requirements.txt
discord.py[voice]==1.7.3
PyNaCl==1.4.0
yt-dlp==2021.12.1
aiohttp==3.7.4
youtube-search==2.1.0
redis
